Moving refresh_mat_view function to dedicated sql_utils file along with run_sql_file.
